Fixed the static linking errors.
