+++ ecs-agent/daemonimages/csidriver/driver/mount.go
@@ -20,14 +20,6 @@ limitations under the License.
 package driver
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
-
-	diskapi "github.com/kubernetes-csi/csi-proxy/client/api/disk/v1"
-	diskclient "github.com/kubernetes-csi/csi-proxy/client/groups/disk/v1"
-
 	"github.com/aws/amazon-ecs-agent/ecs-agent/daemonimages/csidriver/mounter"
 	mountutils "k8s.io/mount-utils"
 )
@@ -69,43 +61,4 @@ func newNodeMounter() (Mounter, error) {
 		return nil, err
 	}
 	return &NodeMounter{safeMounter}, nil
-}
-
-// DeviceIdentifier is for mocking os io functions used for the driver to
-// identify an EBS volume's corresponding device (in Linux, the path under
-// /dev; in Windows, the volume number) so that it can mount it. For volumes
-// already mounted, see GetDeviceNameFromMount.
-// https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html#identify-nvme-ebs-device
-type DeviceIdentifier interface {
-	Lstat(name string) (os.FileInfo, error)
-	EvalSymlinks(path string) (string, error)
-	ListDiskIDs() (map[uint32]*diskapi.DiskIDs, error)
-}
-
-type nodeDeviceIdentifier struct{}
-
-func newNodeDeviceIdentifier() DeviceIdentifier {
-	return &nodeDeviceIdentifier{}
-}
-
-func (i *nodeDeviceIdentifier) Lstat(name string) (os.FileInfo, error) {
-	return os.Lstat(name)
-}
-
-func (i *nodeDeviceIdentifier) EvalSymlinks(path string) (string, error) {
-	return filepath.EvalSymlinks(path)
-}
-
-func (i *nodeDeviceIdentifier) ListDiskIDs() (map[uint32]*diskapi.DiskIDs, error) {
-	diskClient, err := diskclient.NewClient()
-	if err != nil {
-		return nil, fmt.Errorf("error creating csi-proxy disk client: %q", err)
-	}
-	defer diskClient.Close()
-
-	response, err := diskClient.ListDiskIDs(context.TODO(), &diskapi.ListDiskIDsRequest{})
-	if err != nil {
-		return nil, fmt.Errorf("error listing disk ids: %q", err)
-	}
-	return response.GetDiskIDs(), nil
 }
