+++ ecs-agent/daemonimages/csidriver/driver/mount_linux.go
@@ -25,6 +25,7 @@ package driver
 import (
 	"fmt"
 	"os"
+	"path/filepath"
 	"strconv"
 	"strings"
 
@@ -163,4 +164,28 @@ func (m *NodeMounter) Unstage(path string) error {
 
 func (m *NodeMounter) NewResizeFs() (Resizefs, error) {
 	return mountutils.NewResizeFs(m.Exec), nil
+}
+
+// DeviceIdentifier is for mocking os io functions used for the driver to
+// identify an EBS volume's corresponding device (in Linux, the path under
+// /dev; in Windows, the volume number) so that it can mount it. For volumes
+// already mounted, see GetDeviceNameFromMount.
+// https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html#identify-nvme-ebs-device
+type DeviceIdentifier interface {
+	Lstat(name string) (os.FileInfo, error)
+	EvalSymlinks(path string) (string, error)
+}
+
+type nodeDeviceIdentifier struct{}
+
+func newNodeDeviceIdentifier() DeviceIdentifier {
+	return &nodeDeviceIdentifier{}
+}
+
+func (i *nodeDeviceIdentifier) Lstat(name string) (os.FileInfo, error) {
+	return os.Lstat(name)
+}
+
+func (i *nodeDeviceIdentifier) EvalSymlinks(path string) (string, error) {
+	return filepath.EvalSymlinks(path)
 }
